% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_theme.R
\name{ggsave_fm}
\alias{ggsave_fm}
\title{Fors Marsh themed plot saving}
\usage{
ggsave_fm(filename, ...)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{...}{Other arguments passed on to \code{ggsave} and the graphics device
function, as specified by device.}
}
\description{
Saving method appropriate for Fors Marsh themed graphics.
}
\details{
Note that the \code{width} and \code{height} arguments cannot be passed to \code{ggsave}.
These values are fixed by \code{ggsave_fm} and will produce an error if
specified by the user.
}
