% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_theme.R
\name{theme_fm}
\alias{theme_fm}
\alias{scale_linetype_fm}
\alias{scale_shape_fm}
\alias{scale_fillshape_fm}
\alias{scale_fillbar_fm}
\alias{scale_linebar_fm}
\alias{geom_col_fm}
\alias{geom_point_fm}
\title{Fors Marsh ggplot2 theme}
\usage{
theme_fm(fontpath = "C:/Windows/Fonts/", ...)

scale_linetype_fm(...)

scale_shape_fm(...)

scale_fillshape_fm(...)

scale_fillbar_fm(...)

scale_linebar_fm(...)

geom_col_fm(...)

geom_point_fm(...)
}
\arguments{
\item{fontpath}{Path to Berlingske Sans fonts.}

\item{...}{Passes arguments to other methods.}
}
\description{
A \code{ggplot} theme that uses Fors Marsh style guidelines.
}
\details{
Fors Marsh's brand guidelines use fonts, color, linestyles, and symbols as
distinctive visual markers of the Seriously Human Fors Marsh brand in
graphics depicting data analysis results.

\emph{FMtheme} includes the \code{theme_fm} that applies the background style for all
graphics and fonts. Aesthetics such as colors, shapes and linetypes are
implemented in separate functions to be used as \code{scale} functions.
These scale functions, and several implementations of \code{geom}s, are discussed
in separate sections below.

Thematic elements that currently do not have a specific implementation in
Fors Marsh theme inheriting from \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw}}.
The \emph{FMtheme} package supports most line-, bar-, and dot-based graphics.
Our recommendations on how to implement each is discussed below.
\subsection{Line Graphs}{

Line-based graphics tend to use \code{\link[ggplot2:geom_path]{ggplot2::geom_line}} and the functions in
\emph{FMtheme} are intended to work with the \emph{linetype} aesthetic for this geom.

We recommend keeping all lines black in color and using \code{scale_linetype_fm}
to differentiate different trendlines when used as an aesthetic in the
same plot. \code{scale_linetype_fm} allows up to five different line patterns in
the same plot.
}

\subsection{Bar Graphs}{

Bar-based graphics tend to use \code{\link[ggplot2:geom_bar]{ggplot2::geom_col}} and the functions in
\emph{FMtheme} are intended to work with both the \emph{linetype} and \emph{fill}
aesthetics for these geoms.

We recommend using a combination of \code{scale_linebar_fm} and \code{scale_fillbar_fm}
to give aesthetics to both the \emph{linetype} and \emph{fill} when
differentiating between trends in the same plot.
\code{scale_linebar_fm} and \code{scale_fillbar_fm} allow up to five different line
patterns and bar colors in the same plot. We also recommend using
\code{geom_col_fm} when creating a bar graph to give the \emph{linetype} and
\emph{color} aesthetics the correct defaults.

It is important to always use both \code{scale_linebar_fm} and \code{scale_fillbar_fm}
when constructing bar graphs as, when used alone, these aesthetics can fail
to be 508 compliant as the \emph{fill} colors are not designed to be
distinguishable. We also recommend using \code{\link[ggplot2:position_dodge]{ggplot2::position_dodge}}
when using multiple trends as opposed to other position methods.

When plotting a histogram or other similar single trend, consider using
only black \emph{fill} and black \emph{color} bars with solid \emph{linetype} instead of
these scales. These scales are mostly useful for depicting multiple trends.
}

\subsection{Dot Graphs}{

Dot-based graphics use \code{\link[ggplot2:geom_point]{ggplot2::geom_point}} and the functions in \emph{FMtheme}
are intended to work with both the \emph{shape} and \emph{fill} aesthetics for this
geom.

We recommend using a combination of \code{scale_shape_fm} and \code{scale_fillshape_fm}
to give aesthetics to both the \emph{shape} and \emph{fill} when
differentiating between trends in the same plot.
\code{scale_shape_fm} and \code{scale_fillshape_fm} allow up to five different shapes
and shape colors in the same plot. We also recommend using
\code{geom_shape_fm} when creating a bar graph to give the \emph{color} and \emph{size}
aesthetics the correct defaults.

It is important to always use both \code{scale_shape_fm} and \code{scale_fillsjape_fm}
when constructing dot graphs as, when used alone, these aesthetics can fail
to be 508 compliant as the \emph{fill} colors are not designed to be
distinguishable.

When plotting a scatterplot or other similar single trend, consider using
only black \emph{fill} and number 16/filled in circle \emph{shape} points with
black \emph{color} instead of these scales. These scales are mostly useful for
depicting multiple trends.
}
}
\section{Key Considerations}{


There are a number of strong suggestions for how to use the functions in
\emph{FMtheme} when generating graphs. If you want to use any of the below these
graph features or change defaults to the theme, scales, or geoms defined
in the provided functions, first consult the Creative Craft Center.

These suggestions are binned together in conceptual categories below.
\subsection{Don't use \code{facet_grid} or \code{facet_wrap}}{

Broadly, do not use \code{facet_grid} or \code{facet_wrap}. The use of faceting
affects the proportionality of the graphs and can affect 508 compliance in
unexpected ways.

As opposed to generating graphs using faceting, create separate graphs
by the variables on which you were intending to facet and combine them as
separate graphics in the final document in a way that can facilitate
comparing them.
}

\subsection{Don't add more aesthetics}{

Do not add colors, symbols, or other aesthetics to the lines, bars, or
dot graphs.  This includes labels on data points in the graphic.
The design of each graph has been customized to be branded, 508 compliant,
easy to read for clients and stakeholders, and easy to use for R users with
\emph{ggplot2} familiarity.

Changing visual design aspects, adding text to the graph region, and other
elements of the graph may make the graph fail to be branded, 508 compliant,
or easy to read.
}

\subsection{Don't add titles or captions}{

Do not add graph titles, subtitles, captions, and other notes that are added
to graphs. Any such title or note could be added to the document in which
the graphic is to be included and is probably a lot easier to edit in the
document than as embedded in the graph.

Axis titles and legend entries do not apply to this recommendation and
clear, concise labels for aesthetics and axes can be useful for readers.
}

\subsection{Save with \code{ggsave_fm}}{

\code{\link{ggsave_fm}} is designed to retain the proper proportionality for any
graphic produced. Using \code{\link[ggplot2:ggsave]{ggplot2::ggsave}} does not respect the graph's
intended proportionality and may make it 508 uncompliant as a result.
}
}

\examples{
require(ggplot2)
require(datasets)

## Example Bar Chart
mtcars |>
  aggregate(mpg ~ vs + am, mean, data = _) |>
  transform(am = as.factor(am), vs = as.factor(vs)) |>
  ggplot(aes(vs, mpg, linetype = am, fill = am))  +
  scale_linebar_fm() + scale_fillbar_fm() +
  geom_col_fm(position = position_dodge()) + theme_fm()

## Example Line Plot
mtcars |>
  aggregate(mpg ~ gear + cyl, mean, data = _) |>
  transform(gear = as.factor(gear))|>
  ggplot(aes(cyl, mpg, linetype = gear, group = gear)) +
  scale_linetype_fm() + geom_line() + theme_fm()

## Example Dot Plot
mtcars |>
  transform(gear = as.factor(gear)) |>
  ggplot(aes(qsec, mpg, shape = gear, fill = gear)) +
  geom_point_fm() + theme_fm() +
  scale_fillshape_fm() + scale_shape_fm()

}
